+++ src/utils/hooks/useCountryOptions.tsx
@@ -1,3 +1,4 @@
+import { getCountryUrls } from 'story-utils/getCountryUrls';
 import { z } from 'zod';
 
 import {
@@ -30,15 +31,6 @@ const schema = z.array(
     zipRegex: z.string().nullable(),
   }),
 );
-
-export const getCountryUrls = () => {
-  const dashboardUrl = import.meta.env.STORYBOOK_ZONOS_DASHBOARD_URL || null;
-  if (!dashboardUrl) {
-    // eslint-disable-next-line no-console
-    console.error('Missing environment variable STORYBOOK_ZONOS_DASHBOARD_URL');
-  }
-  return { dashboardUrl };
-};
 
 export const useCountryOptions = <TCountryCode extends string>(
   dashboardUrl?: string,
